% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19shiny-package.R
\docType{package}
\name{covid19shiny-package}
\alias{covid19shiny}
\alias{covid19shiny-package}
\title{covid19shiny: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Helen Evangelina \email{heva0001@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
