% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19total.R
\docType{data}
\name{covid19total}
\alias{covid19total}
\title{COVID-19 total dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 183 rows and 9 columns.
}
\source{
The data below comes from \code{tidycovid19} package created by Joachim Gassen, which was originally collected and cleaned by Johns Hopkins University, joined with \code{countries.csv} which was taken from Google Developer.
}
\usage{
covid19total
}
\description{
This dataset contains the summary of COVID-19 data from 183 countries taken from the \code{tiidycovid19} package, alongside with their longitude and latitude.
}
\keyword{datasets}
