% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtable.R
\name{covidtable}
\alias{covidtable}
\title{COVID-19 table function}
\usage{
covidtable(data)
}
\arguments{
\item{data}{the dataset wanting to be put into the datatable}
}
\value{
A daily COVID-19 plot of daily cases, deaths or recovered based on the input
}
\description{
This function creates a table by using the \code{DT} package
}
