% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Running the Shiny Application}
\usage{
launch_app()
}
\value{
Shinydashboard illustrating the state of COVID-19 cases around the world, with user interactions provided
}
\description{
This function will runs the COVID-19 Shinydashboard which presents shinydashboard of the state of COVID-19 in the world
}
