% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worlshp.R
\docType{data}
\name{worldshp}
\alias{worldshp}
\title{worldshp}
\format{
A shp file
}
\source{
thematicmapping.org \url{http://thematicmapping.org/downloads/world_borders.php}
}
\usage{
worldshp
}
\description{
This dataset contains the world shape file with longitude and latitude variables needed to produce a map
}
\keyword{datasets}
