% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19shiny-package.R
\docType{package}
\name{covid19shiny-package}
\alias{covid19shiny}
\alias{covid19shiny-package}
\title{covid19shiny: Shiny Dashboard of Coronavirus State Worldwide}
\description{
Presenting an overview of Coronavirus state in the world in a Shinydashboard by providing functions to launch the application and simplify as well as produce the contents inside the Shinydashboard.
}
\author{
\strong{Maintainer}: Helen Evangelina \email{heva0001@student.monash.edu}

}
\keyword{internal}
