% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19total.R
\docType{data}
\name{covid19total}
\alias{covid19total}
\title{covid19total}
\format{
a dataset with
\itemize{
\item \emph{add variables}
}
}
\source{
tidycovid19 package on Github
}
\usage{
covid19total
}
\description{
This data set contains the total cases, deaths, recovered by country of COVID-19
}
\keyword{datasets}
