% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19wrangled.R
\docType{data}
\name{covid19wrangled}
\alias{covid19wrangled}
\title{covid19wrangled dataset}
\format{
a dataset with
\itemize{
\item \emph{add variables}
}
}
\source{
tidycovid19 Github
}
\usage{
covid19wrangled
}
\description{
This data set contains
}
\keyword{datasets}
