% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19wrangled.R
\docType{data}
\name{covid19wrangled}
\alias{covid19wrangled}
\title{COVID-19 dataset wrangled}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48861 rows and 7 columns.
}
\source{
The data below comes from tidycovid19 package created by Joachim Gassen, which was originally collected and cleaned by Johns Hopkins University.
}
\usage{
covid19wrangled
}
\description{
This dataset contains COVID-19 data from 183 countries taken from the \code{tiidycovid19} package with only country, date, confirmed, deaths, recovered, region, population variables selected and omitting the NA values.
}
\keyword{datasets}
