% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtable.R
\name{covidtable}
\alias{covidtable}
\title{COVID-19 table function}
\usage{
covidtable(data)
}
\arguments{
\item{data}{the dataset wanted to be put into the table}
}
\value{
A table presenting data based on specified dataset
}
\description{
This function creates a table by using the \code{DT} package
}
\examples{
 \dontrun{covidtable(covid19total)}

}
