% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyplot.R
\name{dailyplot}
\alias{dailyplot}
\title{Covid-19 plot function}
\usage{
dailyplot(input_function, input_name, variable_name)
}
\arguments{
\item{a}{}
}
\value{
A daily covid-19 plot of daily cases, deaths or recovered based on the input
}
\description{
Creates daily COVID-19 plot based on the input
}
