% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectInput01.R
\name{selectInput01}
\alias{selectInput01}
\title{SelectInput for the shiny app}
\usage{
selectInput01(id, choices)
}
\arguments{
\item{A}{function to add selectInput on the shiny app}
}
\description{
SelectInput for the shiny app
}
