% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worlshp.R
\docType{data}
\name{worldshp}
\alias{worldshp}
\title{worldshp}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 246 rows and 11 columns.
}
\source{
aaa
}
\usage{
worldshp
}
\description{
world shape file
}
\keyword{datasets}
